UI Progress (Modal, Table)

+++ app/Http/Controllers/UserController.php
@@ -106,7 +106,13 @@ class UserController extends Controller
    
     public function destroy($id)
     {
-        User::find($id)->delete();
+        $user = User::find($id);
+        if ($user->roles()->where('name', 'superadmin')->exists()) {
+            return redirect()->route('users.index')
+                        ->with('info', $user->name.' cannot be deleted!');
+        } else {
+            User::find($id)->delete();
+        }
         return redirect()->route('users.index')
                         ->with('success','User deleted successfully');
     }
